Base

Types

type CN = Complex Double

Type alias for a complex number.

type XY = (Double, Double)

Type alias for a coordinate representation.

Directed values

data Angular

The representation of a directed value isomorphic either to an angle or to a complex number or coordinates.

Angular isomorphisms

deg :: Angular -> Double

Returns a representation of a directed value as an angle in degrees.

asDeg :: Double -> Angular

Constructs a directed value from an angle given in degrees.

rad :: Angular -> Double

Returns a representation of a directed value as an angle in radians.

asRad :: Double -> Angular

Constructs a directed value from an angle given in radians.

turns :: Angular -> Double

Returns a representation of a directed value as a number of turns.

asTurns :: Double -> Angular

Constructs a directed value from a number of turns.

Linear transformations

type TMatrix = ((Double, Double, Double), (Double, Double, Double))

The representation of a linear transformation matrix.

rotateT :: Double -> TMatrix

The rotation matrix.

reflectT :: Double -> TMatrix

The reflection matrix.

translateT :: CN -> TMatrix

The translation matrix.

scaleT :: Double -> Double -> TMatrix

The scaling matrix.

Classes

Points in affine space

class Affine a where

Class representing points in 2d affine space. Instances may be points, coordinates, complex numbers, 2d-vectors.

Minimal complete definition

(asCmp | asCoord), (cmp | coord)

Methods

asCmp :: CN -> a

Constructs an affine point from a complex number.

cmp :: a -> CN

Returns a representation of an afine point as a complex number.

asCoord :: XY -> a

Constructs an affine point from a coordinates.

coord :: a -> XY

Returns a representation of an afine point as a coordinate.

getX :: a -> Double

Returns x-coordinate of an afine point.

getY :: a -> Double

Returns y-coordinate of an afine point.

Predicates

isOrthogonal :: (Affine a, Affine b) => a -> b -> Bool

Returns True if two points represent orthogonal vectors.

isCollinear :: (Affine a, Affine b) => a -> b -> Bool

Returns True if two points represent collinear vectors.

isOpposite :: (Affine a, Affine b) => a -> b -> Bool

Returns True if two points represent collinear opposite vectors.

isZero :: Affine a => a -> Bool

Returns True if the vector is trivial or a point is equal to the origin.

Vector and point operations

dot :: (Affine a, Affine b) => a -> b -> Double

The dot product of two points.

det :: (Affine a, Affine b) => (a, b) -> Double

Determinant of a matrix, composed of two vector rows. Equivalent to cross.

cross :: (Affine a, Affine b) => a -> b -> Double

Z-component of a cross product of two vectors. Equivalent to det.

norm :: Affine a => a -> Double

The norm of a vector, or a distance from a point to the origin.

distance :: (Affine a, Affine b) => a -> b -> Double

The distance between tho points, or a norm of the difference between two vectors.

angle :: Affine a => a -> Angular

The direction of the vector or a point.

normalize :: Affine a => a -> a

The normalized vector, or a projection of a point on a unit circle.

columns :: Affine a => (a, a) -> (a, a)

The matrix, composed of two vector columns.

Linear transformations

class Trans a where

Class for objects which could be transformed linearly.

Methods

transform :: TMatrix -> a -> a

The general linear transformation.

transformCN :: TMatrix -> CN -> CN

The transformation of a complex number. Used in class implementations.

transformXY :: TMatrix -> XY -> XY

The transformation of coordinates. Used in class implementations.

Fuzzy equality

class AlmostEq a

Type class for values for which equality could be stated with some known tolerance.

Minimal complete definition

(~==)

Fuzzy inequalities

(~<=) :: (AlmostEq a, Ord a) => a -> a -> Bool infix 4

The less or almost equal relation.

(~>=) :: (AlmostEq a, Ord a) => a -> a -> Bool infix 4

The greater or almost equal relation.