Circle

Types and classes

data Circle

Represents a circle with given center, passing through given point.

Constructors

Circle

Fields

CN

center,

CN

starting point,

Double

orientation poitive -- CW, negative -- CCW.

class Curve c => IsCircle c where

Class for circle and decorated circle

Minimal complete definition

center, radius, orientation, phaseShift

Methods

center :: c -> CN

Center of the circle.

radius :: c -> Double

Radius of the circle

orientation :: c -> Double

Orientation of the circle

phaseShift :: c -> Double

The angle of the starting point.

radiusVector :: c -> Double -> XY

The radius-vector for a given parameter

Constructors

trivialCircle :: Circle

The trivial circle with zero radius.

mkCircle :: CN -> CN -> Circle

The constructor for a circle with given center and starting point.

mkCircleRC :: Double -> CN -> Circle

The constructor for a circle with given center and radius. The starting point is located at zero angle.