Decorations

Classes

class Decor a where

The class of objects which could have decorations.

Minimal complete definition

Nothing

Methods

options :: a -> Options

Get the decoration data

setOptions :: Options -> a -> a

Set the decoration data

labelDefaults :: a -> Labeling

Default labeling settings

styleDefaults :: a -> Style

Default styling settings

Records and Types

type Options = (Labeling, Style)

The alias for a tuple of decoration records

data Labeling

Record representing labeling parameters

Constructors

Labeling

Fields

getLabel :: Last String

getLabelCorner :: Last (Int, Int)

getLabelOffset :: Last XY

getLabelPosition :: Last CN

getLabelAngle :: Last Angular

data Style

Record representing styling parameters

Constructors

Style

Fields

getStroke :: Last String

getFill :: Last String

getDashing :: Last String

getStrokeWidth :: Last String

newtype Decorated a

The transparent decoration wrapper for geometric objects. Inherits all properties of embedded object.

Constructors

Decorated (Options, a)

fromDecorated :: Decorated a -> a

The selector for the embedded object.

Decorators

(#:) :: Decor a => a -> Decorator a -> Decorated a infixl 5

The infix operator for decorator apprication.

>>> aPoint # at (4, 5) #: label "A"
A:<Point (4.0, 5.0)>

>>> segment (4,5) (6,9) #: "s" <> dotted <> white
s:<Segment (4.0,5.0) (6.0,9.0)>

stroke :: Decor a => String -> Decorator a

The stroke color decorator.

white :: Decor a => Decorator a

The decorator for white lines.

fill :: Decor a => String -> Decorator a

The fill color decorator.

width :: Decor a => String -> Decorator a

The stroke-width decorator.

thin :: Decor a => Decorator a

The decorator for thin lines.

dotted :: Decor a => Decorator a

The decorator for dotted lines.

dashed :: Decor a => Decorator a

The decorator for dashed lines.

label :: Decor a => String -> Decorator a

The decorator for labeling objects. Could be used as overloaded string.

>>> aPoint # at (4, 5) #: "A"
A:<Point (4, 5)>

loffs :: Decor a => XY -> Decorator a

The decorator for label offset.

lpos :: (Affine p, Decor a) => p -> Decorator a

The decorator for label position.

lparam :: (Curve a, Decor a) => Double -> Decorator a

The decorator for setting label on a curve at a given parameter value.

Functions

labelText :: Decor a => a -> Maybe String

Returns label text of a decorated object.

labelPosition :: Decor a => a -> CN

Returns label position for a decorated object.

labelOffset :: Decor a => a -> XY

Returns label offset against the label position.

labelCorner :: Decor a => a -> (Int, Int)

Returns label corner used as an anchor of a text in SVG.