Figures

Constructors for geometric objects

Point constructors

origin :: Point

The origin point

aPoint :: Point

The template for a point.

aLabel :: Label

The template for a label.

point :: XY -> Point

The constructor for a point with given coordinates.

point' :: Affine a => a -> Point

The constructor for a point with given affine coordinates.

pointOn :: Curve a => a -> Double -> Point

The point on a given curve.

projectOn :: (Curve c, Affine p) => p -> c -> Point

Returns a point on the curve which is s normal projection of the given point on the curve.

intersectionPoints :: (Intersections a b, Curve a, Curve b) => a -> b -> [Point]

Returns a list of intersection points as geomtric figures.

Line constructors

aLine :: Line

The template for a line.

aRay :: Line

The template for a ray.

aSegment :: Line

The template for a segment.

oX :: Line

The x-axis

oY :: Line

The y-axis

line :: XY -> XY -> Line

line' :: (Affine a1, Affine a2) => a1 -> a2 -> Line

ray :: XY -> XY -> Line

ray' :: (Affine a1, Affine a2) => a1 -> a2 -> Line

segment :: XY -> XY -> Line

segment' :: (Affine a1, Affine a2) => a1 -> a2 -> Line

Angle constructors

anAngle :: Angular -> Angle

The template for an angle with given value

angleBetween :: (IsLine l1, IsLine l2) => l1 -> l2 -> Angle

Returns the angle equal to the angle between thwo lines, located on the first one.

adjacent :: Angle -> Angle

vertical :: Angle -> Angle

complementary :: Angle -> Angle

bisectrisse :: IsAngle a => a -> Line

Polygon constructors

aTriangle :: Triangle

The template for a triangle.

triangle2a :: Angular -> Angular -> Polygon

Returns a triangle with base 1 and two given angles.

aSquare :: Polygon

The template for a square.

aRectangle :: Double -> Double -> Polygon

The template for a Rectangle.

parametricPoly :: (Double -> XY) -> [Double] -> Polygon

Constructs a parametric graph as a Polyline.

polarPoly :: (Double -> Double) -> [Double] -> Polygon

Constructs a polar graph as a Polyline.

regularPoly :: Int -> Polygon

Constructs a regular polygon with given number of sides, enscribed in a unit circle.

Circle constructors

aCircle :: Circle

The template for a circle.

circle :: Double -> XY -> Circle

circle' :: Affine a => Double -> a -> Circle

Modificators

at :: Figure a => XY -> a -> a

at' :: (Affine p, Figure a) => p -> a -> a

along :: (Figure a, Affine a) => Double -> a -> a

along' :: (Figure f, Affine v, Affine f) => v -> f -> f

through :: IsLine l => XY -> l -> l

A coordinated version of through.

through' :: (Affine p, IsLine l) => p -> l -> l

Turns and extends the line so that it passes through a given point.

translate :: Trans a => XY -> a -> a

scaleAt :: Trans a => XY -> Double -> a -> a

scaleXAt :: Trans a => XY -> Double -> a -> a

scaleYAt :: Trans a => XY -> Double -> a -> a

on :: (Figure a, Affine a, Curve c) => c -> Double -> a -> a

Locates an affine object on a given curve agt given parameter and aligns it along a tangent to a curve.

normalTo :: (Curve c, IsLine l) => c -> l -> l

Turns the line so that it becomes normal to a given curve, pointing towards a curve.