Geometry

Reexports modules

module Base

module Point

module Line

module Circle

module Polygon

module Decorations

Main interface

writeSVG :: SVGable a => FilePath -> a -> IO ()

showSVG :: SVGable a => a -> Text

Creates a SVG contents for geometric objects.

(<+>) :: (Groupable a, Groupable b) => a -> b -> Group infixl 5

The appending operator for groupable objects.

group :: Groupable a => [a] -> Group

Returns a group of homogeneous list of objects.

Constructors for affine geometric objects

aPoint :: Point

aLabel :: String -> Decorated Label

aLine :: Line

aRay :: Line

aSegment :: Line

aCircle :: Circle

aSquare :: Polygon

aRectangle :: Double -> Double -> Polygon

aTriangle :: Polygon

Constructors for exact objects

origin :: Point

oX :: Line

oY :: Line

point :: XY -> Point

pointOn :: Curve a => a -> Double -> Point

circle :: Double -> XY -> Circle

line :: XY -> XY -> Line

ray :: XY -> XY -> Line

segment :: XY -> XY -> Line

parametricPoly :: (Double -> XY) -> [Double] -> Polygon

polarPoly :: (Double -> Double) -> [Double] -> Polygon

regularPoly :: Int -> Polygon

Modificators

at :: Figure a => XY -> a -> a

along :: (Figure a, Affine a) => Double -> a -> a

through :: Linear l => XY -> l -> l

A coordinated version of through.

perpendicularTo :: Linear l => Line -> l -> l

Turns the line so that it becomes perpendicular to a given one, pointing towards it.

General versions of conctructors and modifiers

point' :: Affine a => a -> Point

line' :: (Affine a1, Affine a2) => a1 -> a2 -> Line

ray' :: (Affine a1, Affine a2) => a1 -> a2 -> Line

segment' :: (Affine a1, Affine a2) => a1 -> a2 -> Line

circle' :: Affine a => Double -> a -> Circle

at' :: (Affine p, Figure a) => p -> a -> a

along' :: (Figure f, Affine v, Affine f) => v -> f -> f

through' :: (Affine p, Linear l) => p -> l -> l

Turns and extends the line so that it passes through a given point.