Geometry.Circle

Types and classes

data Circle

Represents a circle with given center, radius vector and tangent direction.

class (Trans c, Manifold CN c, Curve c, ClosedCurve c, Figure c) => Circular c where

Class for circle and decorated circle

Methods

center :: c -> CN

Center of the circle.

radius :: c -> Double

Radius of the circle

phaseShift :: c -> Direction

The angle of the starting point.

orientation :: c -> Direction

The orientation of the circle (positive -- CCW).

Constructors

trivialCircle :: Circle

The trivial circle with zero radius.

mkCircle :: Double -> CN -> Circle

The constructor for a circle with given center and radius. The radius-vector has zero angle, and circle is CCW-oriented.