Geometry.Polygon

Documentation

class (Trans p, Manifold CN p) => PiecewiseLinear p where

A class for polylines and polygons.

Minimal complete definition

vertices

Methods

vertices :: p -> [CN]

A list of polyline vertices.

asPolyline :: p -> Polyline

A representation of the instance as a polyline.

verticesNumber :: p -> Int

The number of vertices of a polyline.

segments :: p -> [Segment]

The list of segments of a polyline.

vertex :: p -> Int -> CN

The indexed selector for vertices of a polyline.

side :: p -> Int -> Segment

The indexed selector for sides of a polyline.

isDegenerate :: PiecewiseLinear p => p -> Bool

A predicate. Returns True if any of polyline's segment has zero length.

newtype Polyline

Representation of a polygonal chain as a list of vertices. Vector representation of a polyline is given by the first segment of a polyline. It is used for alignment or creation of polylines.

>>> cmp $ mkTriangle [0, 1:+0, 0:+1]
1.0 :+ 0.0

>>> asCmp (1 :+ 2) :: Triangle
<Triangle (0.0,0.0) (1.0,2.0) (-1.2320508075688772,1.8660254037844393)>

>>> asCmp (1) :: Rectangle
<Rectangle (0.0,0.0) (1.0,0.0) (1.0,1.0) (0.0,1.0)>

Constructors

Polyline [CN]

mkPolyline :: Affine a => [a] -> Polyline

The main polyline constructor.

trivialPolyline :: Polyline

Creates an empty polyline.

newtype Polygon

Representation of a closed polygon as a list of vertices.

Constructors

Polygon [CN]

mkPolygon :: Affine a => [a] -> Polygon

The main polyline constructor.

trivialPolygon :: Polygon

Creates an empty polyline.

closePolyline :: Polyline -> Polygon

Turns a polyline to a polygon with the same vertices.

newtype Triangle

Representation of a triangle as a list of vertices.

Constructors

Triangle [CN]

mkTriangle :: Affine a => [a] -> Triangle

The main triangle constructor.

newtype Rectangle

Representation of a rectangle as a list of vertices.

Constructors

Rectangle [CN]

mkRectangle :: Double -> Double -> Rectangle

The main rectangle constructor, uses two side lengths and guarantees that polygon will have right angles.

boxRectangle :: Figure a => a -> Rectangle

Returns a figures' box as a rectangle.

data Plot

Constructors

Plot (Double -> XY) (Double, Double)

plotManifold :: (Affine a, Manifold a m) => (Double, Double) -> m -> Polyline