Polygon

Documentation

class Curve p => IsPoly p where

A class for polylines and polygons.

Minimal complete definition

vertices, asPolyline

Methods

vertices :: p -> [CN]

asPolyline :: p -> Polyline

verticesNumber :: p -> Int

segments :: p -> [Segment]

vertex :: p -> Int -> CN

side :: p -> Int -> Segment

isDegenerate :: IsPoly p => p -> Bool

A predicate. Returns True if any of polyline's segment has zero length.

newtype Polyline

Representation of a polygonal chain as a list of vertices. Polyline inherits monoidal properties of lists.

Constructors

Polyline [CN]

mkPolyline :: Affine a => [a] -> Polyline

The main polyline constructor.

trivialPolyline :: Polyline

Creates an empty polyline.

newtype Polygon

Representation of a closed polygon as a list of vertices.

Constructors

Polygon [CN]

mkPolygon :: Affine a => [a] -> Polygon

The main polyline constructor.

trivialPolygon :: Polygon

Creates an empty polyline.

closePolyline :: Polyline -> Polygon

Turns a polyline to a polygon with the same vertices.

newtype Triangle

Representation of a triangle as a list of vertices.

Constructors

Triangle [CN]

mkTriangle :: Affine a => [a] -> Triangle

The main triangle constructor.

newtype Rectangle

Representation of a rectangle as a list of vertices.

Constructors

Rectangle [CN]

mkRectangle :: Double -> Double -> Rectangle

The main rectangle constructor, uses two side lengths and guarantees that polygon will have right angles.

boxRectangle :: Figure a => a -> Rectangle