SVG

Description

This module defines representation of all figures as SVG elements. as well as default style settings.

Classes

type Groupable a = (SVGable a, Show a, Trans a, Figure a, Eq a)

Constrain for an object that could be included to a group.

data Group where

The group of inhomogeneous Groupable objects.

Constructors

EmptyFig :: Group

G :: Groupable a => a -> Group

Append :: Group -> Group -> Group

(<+>) :: (Groupable a, Groupable b) => a -> b -> Group infixl 2

The appending operator for groupable objects.

group :: Groupable a => [a] -> Group

Returns a group of homogeneous list of objects.

beside :: (Groupable a, Groupable b) => a -> b -> Group

class SVGable a where

Class for objects that could be represented as SVG elements. Differs from ToElement class in adding parameter context to a main render function toSVG.

Minimal complete definition

Nothing

Methods

toSVG :: a -> SVGContext -> Element

fmtSVG :: a -> Text

type ImageSize = Int

Type alias for explicit image settings

data SVGContext

The record containing image parameters and options.

Constructors

SVGContext

Fields

imageSize :: ImageSize

figureBox :: Rectangle

border :: Double

figureOptions :: Options

Functions

showSVG :: (Figure a, SVGable a) => ImageSize -> a -> Text

Creates a SVG contents for geometric objects.