Base

Types

type CN = Complex Double

Type alias for a complex number.

type XY = (Double, Double)

Type alias for a coordinate representation.

Directed values

newtype Angular

The representation of a directed value isomorphic either to an angle or to a complex number or coordinates.

Constructors

Angular Double

Angular isomorphisms

deg :: Angular -> Double

Returns a representation of a directed value as an angle in degrees.

asDeg :: Double -> Angular

Constructs a directed value from an angle given in degrees.

rad :: Angular -> Double

Returns a representation of a directed value as an angle in radians.

asRad :: Double -> Angular

Constructs a directed value from an angle given in radians.

turns :: Angular -> Double

Returns a representation of a directed value as a number of turns.

asTurns :: Double -> Angular

Constructs a directed value from a number of turns.

Linear transformations

type TMatrix = ((Double, Double, Double), (Double, Double, Double))

The representation of a linear transformation matrix.

rotateT :: Double -> TMatrix

The rotation matrix.

reflectT :: Double -> TMatrix

The reflection matrix.

translateT :: CN -> TMatrix

The translation matrix.

scaleT :: Double -> Double -> TMatrix

The scaling matrix.

Classes

Points in affine space

class Affine a where

Class representing points in 2d affine space. Instances may be points, coordinates, complex numbers, 2d-vectors.

Minimal complete definition

asCmp, cmp | asCoord, coord

Methods

asCmp :: CN -> a

Constructs an affine point from a complex number.

cmp :: a -> CN

Returns a representation of an afine point as a complex number.

asCoord :: XY -> a

Constructs an affine point from a coordinates.

coord :: a -> XY

Returns a representation of an afine point as a coordinate.

getX :: a -> Double

Returns x-coordinate of an afine point.

getY :: a -> Double

Returns y-coordinate of an afine point.

Predicates

isOrthogonal :: (Affine a, Affine b) => a -> b -> Bool

Returns True if two points represent orthogonal vectors.

isCollinear :: (Affine a, Affine b) => a -> b -> Bool

Returns True if two points represent collinear vectors.

isOpposite :: (Affine a, Affine b) => a -> b -> Bool

Returns True if two points represent collinear opposite vectors.

isZero :: Affine a => a -> Bool

Returns True if the vector is trivial or a point is equal to the origin.

Vector and point operations

dot :: (Affine a, Affine b) => a -> b -> Double

The dot product of two points.

det :: (Affine a, Affine b) => (a, b) -> Double

Determinant of a matrix, composed of two vector rows. Equivalent to cross.

cross :: (Affine a, Affine b) => a -> b -> Double

Z-component of a cross product of two vectors. Equivalent to det.

norm :: Affine a => a -> Double

The norm of a vector, or a distance from a point to the origin.

distance :: (Affine a, Affine b) => a -> b -> Double

The distance between tho points, or a norm of the difference between two vectors.

angle :: Affine a => a -> Angular

The direction of the vector or a point.

normalize :: Affine a => a -> a

The normalized vector, or a projection of a point on a unit circle.

columns :: Affine a => (a, a) -> (a, a)

The matrix, composed of two vector columns.

azimuth :: (Affine a, Affine b) => a -> b -> Angular

The ngle between tho two points

Linear transformations

class Trans a where

Class for objects which could be transformed linearly.

Methods

transform :: TMatrix -> a -> a

The general linear transformation.

transformCN :: TMatrix -> CN -> CN

The transformation of a complex number. Used in class implementations.

transformXY :: TMatrix -> XY -> XY

The transformation of coordinates. Used in class implementations.

transformAt :: (Trans a, Affine p) => p -> (a -> a) -> a -> a

The transformation with shifted origin.

transformations

translate :: (Trans a, Affine p) => p -> a -> a

The translation of an object.

superpose :: (Trans a, Affine p1, Affine p2) => p1 -> p2 -> a -> a

The translation leading to a superposition of tho points.

scale :: Trans a => Double -> a -> a

The isotropic scaling of an object.

scaleX :: Trans a => Double -> a -> a

The scaling of an object in x-direction.

scaleY :: Trans a => Double -> a -> a

The scaling of an object in y-direction.

scaleAt :: (Trans a, Affine p) => p -> Double -> a -> a

The isotropic scaling of an object against a given point.

scaleXAt :: (Trans a, Affine p) => p -> Double -> a -> a

The isotropic scaling of an object in x-direction against a given point.

scaleYAt :: (Trans a, Affine p) => p -> Double -> a -> a

The isotropic scaling of an object in y-direction against a given point.

rotate :: Trans a => Angular -> a -> a

The rotation of an object against the origin.

rotateAt :: (Trans a, Affine p) => p -> Angular -> a -> a

The rotation of an object against a given point.

reflect :: Trans a => Angular -> a -> a

The reflection of an object against the direction passing through the origin.

reflectAt :: (Trans a1, Curve a2, Affine a2) => a2 -> a1 -> a1

Curves

class Curve c where

Class representing a curve parameterized by a real number. For finite curves parameter runs from 0 to 1, where 0 is a start and 1 is the end of the curve. The length scale is given by unit function.

All implementations must obay isomorphisms laws:

isFinite c          ==>  (project c . param c) x == x `mod'` 1

not isFinite c      ==>  (project c . param c) x == x

c `isContaining` p  ==>  (param c . project c) p == p

Minimal complete definition

(param | paramMaybe), (project | projectMaybe), (normal | tangent), distanceTo, (location | isContaining, isEnclosing)

Methods

param :: c -> Double -> CN

Returns a point on a curve for given parameter. If parameter value is out of range [0, 1] for a finite curves the closest boundary value is returned. Use paramMaybe for more explicit parameterization.

paramMaybe :: c -> Double -> Maybe CN

Returns a point on a curve for given parameter, or Nothing if parameter runs out of the range defined for a curve.

project :: Affine p => c -> p -> Double

Returns a parameter, corresponding to a normal point projection on a curve. if normal projection doesn't exist returns the parameter of the closest point. Use projectMaybe for more explicit projection.

projectMaybe :: Affine p => c -> p -> Maybe Double

Returns a parameter for a point on a curve, or nothing if parameter could not be found.

unit :: c -> Double

The internal length unit of the curve, which maps the parameter to the length of the curve.

distanceTo :: Affine p => p -> c -> Double

The distance between a curve and a point.

tangent :: c -> Double -> Angular

The tangent direction for a given parameter on the curve.

normal :: c -> Double -> Angular

The normal direction for a given parameter on the curve.

isClosed :: c -> Bool

Is set True if the curve is closed.

isFinite :: c -> Bool

Is set True if the curve is finite.

location :: Affine p => p -> c -> PointLocation

Returns the location of a point with respect to the curve.

isContaining :: Affine p => c -> p -> Bool

Returns True if point belongs to the curve.

isEnclosing :: Affine p => c -> p -> Bool

Returns True if point belongs to the area bound by closed curve.

data PointLocation

The type representing the relation 'belongs to' between a point and a curve.

Constructors

Inside

Outside

OnCurve

(->@) :: (Curve c, Affine p) => p -> c -> Double infix 8

Operator for project with flipped arguments:

p ->@ c  ==  project c p

(->@?) :: (Curve c, Affine p) => p -> c -> Maybe Double infix 8

Operator for projectMaybe with flipped arguments:

p ->@? c  ==  projectMaybe c p

(@->) :: Curve c => c -> Double -> CN infix 8

Operator for param

(@->?) :: Curve c => c -> Double -> Maybe CN infix 8

Operator for paramMaybe

start :: Curve a => a -> CN

The starting point on the curve.

paramL :: Curve c => c -> Double -> CN

Point on the curve, parameterized by length.

projectL :: (Affine p, Curve c) => c -> p -> Double

Projection on a curve, parameterized by length.

Intersections of curves

class (Curve a, Curve b) => Intersections a b where

Class provides intersections function returning a list (possible empty) of intersection points (co-dimension 1).

Methods

intersections :: a -> b -> [CN]

Figures

class (Eq a, Trans a) => Figure a where

Class representing the interface for a figure on a chart

Minimal complete definition

isTrivial, refPoint, box

Methods

isTrivial :: a -> Bool

Returns True is figure is trivial in a certain sence.

isNontrivial :: a -> Bool

Returns True is figure is not trivial in a certain sence.

refPoint :: a -> CN

Returns a refference point using for superposing and figure locataion.

isSimilar :: a -> a -> Bool

The geometric similarity relation.

box :: a -> Box

pointBox :: Affine p => p -> Box

height :: Figure f => f -> Double

width :: Figure f => f -> Double

corner :: Figure a => a -> ((Complex Double, Complex Double), (Complex Double, Complex Double))

left :: (a, b) -> a

right :: (a, b) -> b

lower :: (a, b) -> b

upper :: (a, b) -> a

Fuzzy equality

class AlmostEq a

Type class for values for which equality could be stated with some known tolerance.

Minimal complete definition

(~==)

Fuzzy inequalities

(~<=) :: (AlmostEq a, Ord a) => a -> a -> Bool infix 4

The less or almost equal relation.

(~>=) :: (AlmostEq a, Ord a) => a -> a -> Bool infix 4

The greater or almost equal relation.

(~==) :: AlmostEq a => a -> a -> Bool infix 4

The equality operator.

Misc

(#) :: a -> (a -> c) -> c infixl 5

Flipped application used for right chaining of transformations.

Orphan instances

Bounded Double

Instance details

Methods

minBound :: Double

maxBound :: Double