Base

Types

type CN = Complex Double

Type alias for a complex number.

type XY = (Double, Double)

Type alias for a coordinate representation.

Directed values

newtype Direction

The representation of a directed value isomorphic either to an angle or to a complex number or coordinates.

Constructors

Direction Double

Direction isomorphisms

deg :: Direction -> Double

Returns a representation of a directed value as an angle in degrees.

asDeg :: Double -> Direction

Constructs a directed value from an angle given in degrees.

rad :: Direction -> Double

Returns a representation of a directed value as an angle in radians.

asRad :: Double -> Direction

Constructs a directed value from an angle given in radians.

turns :: Direction -> Double

Returns a representation of a directed value as a number of turns.

asTurns :: Double -> Direction

Constructs a directed value from a number of turns.

Linear transformations

type TMatrix = ((Double, Double, Double), (Double, Double, Double))

The representation of a linear transformation matrix.

transformOrientation :: TMatrix -> Double

Returns 1 if transformation preserves curve orientation, and -1 otherwise.

Classes

Points in affine space

class Affine a where

Class representing points in 2d affine space. Instances may be points, coordinates, complex numbers, 2d-vectors.

Minimal complete definition

asCmp, cmp | asCoord, coord

Methods

asCmp :: CN -> a

Constructs an affine point from a complex number.

cmp :: a -> CN

Returns a representation of an afine point as a complex number.

asCoord :: XY -> a

Constructs an affine point from a coordinates.

coord :: a -> XY

Returns a representation of an afine point as a coordinate.

getX :: a -> Double

Returns x-coordinate of an afine point.

getY :: a -> Double

Returns y-coordinate of an afine point.

roundUp :: Affine a => Double -> a -> a

Predicates

isOrthogonal :: (Affine a, Affine b) => a -> b -> Bool

Returns True if two points represent orthogonal vectors.

isCollinear :: (Affine a, Affine b) => a -> b -> Bool

Returns True if two points represent collinear vectors.

isOpposite :: (Affine a, Affine b) => a -> b -> Bool

Returns True if two points represent collinear opposite vectors.

isZero :: Affine a => a -> Bool

Returns True if the vector is trivial or a point is equal to the origin.

Vector and point operations

dot :: (Affine a, Affine b) => a -> b -> Double

The dot product of two points.

det :: (Affine a, Affine b) => (a, b) -> Double

Determinant of a matrix, composed of two vector rows.

cross :: (Affine a, Affine b) => a -> b -> Double

Z-component of a cross product of two vectors.

norm :: Affine a => a -> Double

The norm of a vector, or a distance from a point to the origin.

distance :: (Affine a, Affine b) => a -> b -> Double

The distance between tho points, or a norm of the difference between two vectors.

angle :: Affine a => a -> Direction

The direction of the vector or a point.

normalize :: Affine a => a -> a

The normalized vector, or a projection of a point on a unit circle.

columns :: Affine a => (a, a) -> (a, a)

The matrix, composed of two vector columns.

azimuth :: (Affine a, Affine b) => a -> b -> Direction

The ngle between tho two points

Linear transformations

class Trans a where

Class for objects which could be transformed linearly.

Methods

transform :: TMatrix -> a -> a

The general linear transformation.

transformCN :: TMatrix -> CN -> CN

The transformation of a complex number. Used in class implementations.

transformXY :: TMatrix -> XY -> XY

The transformation of coordinates. Used in class implementations.

transformAt :: (Trans a, Affine p) => p -> (a -> a) -> a -> a

The transformation with shifted origin.

transformations

translate' :: (Trans a, Affine p) => p -> a -> a

The translation of an object.

superpose :: (Trans a, Affine p1, Affine p2) => p1 -> p2 -> a -> a

The translation leading to a superposition of tho points.

scale :: Trans a => Double -> a -> a

The isotropic scaling of an object.

scaleX :: Trans a => Double -> a -> a

The scaling of an object in x-direction.

scaleY :: Trans a => Double -> a -> a

The scaling of an object in y-direction.

scaleAt' :: (Trans a, Affine p) => p -> Double -> a -> a

The isotropic scaling of an object against a given point.

scaleXAt' :: (Trans a, Affine p) => p -> Double -> a -> a

The isotropic scaling of an object in x-direction against a given point.

scaleYAt' :: (Trans a, Affine p) => p -> Double -> a -> a

The isotropic scaling of an object in y-direction against a given point.

rotate :: Trans a => Direction -> a -> a

The rotation of an object against the origin.

rotateAt' :: (Trans a, Affine p) => p -> Direction -> a -> a

The rotation of an object against a given point.

reflect :: Trans a => Direction -> a -> a

The reflection of an object against the direction passing through the origin.

reflectAt :: (Trans a1, Manifold a2, Affine a2) => a2 -> a1 -> a1

Curves

class Manifold m where

Minimal complete definition

param, project, (isContaining | paramMaybe, projectMaybe)

Methods

bounds :: m -> [Double]

Returns bounds for a parameter of a manifold. [a, b] for finite, [a] for semibound manifolds (a is lower bound), [] for unbound manifolds

paramMaybe :: m -> Double -> Maybe CN

Returns a point on a manifold for given parameter, or Nothing if parameter runs out of the range defined for a manifold.

projectMaybe :: Affine p => m -> p -> Maybe Double

Returns a parameter for a point on a manifold, or nothing if parameter could not be found.

param :: m -> Double -> CN

Returns a point on a manifold for a given parameter. If parameter value is out of range [0, 1] for a finite manifolds the closest boundary value is returned.

project :: Affine p => m -> p -> Double

Returns a parameter, corresponding to a normal point projection on a manifold. if normal projection doesn't exist returns the parameter of the closest point.

isContaining :: Affine p => m -> p -> Bool

Returns True if point belongs to the manifold.

unit :: m -> Double

The internal length unit of the curve, which maps the parameter to the length of the curve.

(->@) :: (Manifold m, Affine p) => p -> m -> Double infix 8

Operator for project with flipped arguments:

p ->@ c  ==  project c p

(->@?) :: (Manifold m, Affine p) => p -> m -> Maybe Double infix 8

Operator for projectMaybe with flipped arguments:

p ->@? c  ==  projectMaybe c p

(@->) :: Manifold m => m -> Double -> CN infix 8

Operator for param

(@->?) :: Manifold m => m -> Double -> Maybe CN infix 8

Operator for paramMaybe

start :: Manifold m => m -> CN

The starting point on the curve.

paramL :: Manifold m => m -> Double -> CN

Point on the curve, parameterized by length.

projectL :: (Affine p, Manifold m) => m -> p -> Double

Projection on a curve, parameterized by length.

distanceTo :: (Manifold m, Affine p) => p -> m -> Double

The distance between a curve and a point.

class (Figure c, Trans c, Manifold c) => Curve c where

Class representing a curve parameterized by a real number. For finite curves parameter runs from 0 to 1, where 0 is a start and 1 is the end of the curve. The length scale is given by unit function.

All implementations must obay isomorphisms laws:

isFinite c          ==>  (project c . param c) x == x `mod'` 1

not isFinite c      ==>  (project c . param c) x == x

c `isContaining` p  ==>  (param c . project c) p == p

Minimal complete definition

normal | tangent

Methods

tangent :: c -> Double -> Direction

The tangent direction for a given parameter on the curve.

normal :: c -> Double -> Direction

The normal direction for a given parameter on the curve.

class Curve c => ClosedCurve c where

Class representing a closed region

Minimal complete definition

location | isEnclosing

Methods

location :: Affine p => c -> p -> PointLocation

Returns the location of a point with respect to the region.

isEnclosing :: Affine p => c -> p -> Bool

Returns True if point belongs to the region.

data PointLocation

The type representing the relation 'belongs to' between a point and a curve.

Constructors

Inside

Outside

OnCurve

Figures

class Trans a => Figure a where

Class representing the interface for a figure on a chart

Minimal complete definition

isTrivial, refPoint, box

Methods

isTrivial :: a -> Bool

Returns True is figure is trivial in a certain sence.

isNontrivial :: a -> Bool

Returns True is figure is not trivial in a certain sence.

refPoint :: a -> CN

Returns a refference point using for superposing and figure locataion.

box :: a -> Box

type Box = ((Min Double, Min Double), (Max Double, Max Double))

pointBox :: Affine p => p -> Box

figureHeight :: Figure f => f -> Double

figureWidth :: Figure f => f -> Double

corner :: Figure a => a -> ((Complex Double, Complex Double), (Complex Double, Complex Double))

left :: (a, b) -> a

right :: (a, b) -> b

lower :: (a, b) -> b

upper :: (a, b) -> a

Fuzzy equality

class AlmostEq a

Type class for values for which equality could be stated with some known tolerance.

Minimal complete definition

(~==)

Fuzzy inequalities

(~<=) :: (AlmostEq a, Ord a) => a -> a -> Bool infix 4

The less or almost equal relation.

(~>=) :: (AlmostEq a, Ord a) => a -> a -> Bool infix 4

The greater or almost equal relation.

(~==) :: AlmostEq a => a -> a -> Bool infix 4

The equality operator.

Misc

(#) :: a -> (a -> c) -> c infixl 5

Flipped application used for right chaining of transformations.

(#!) :: t -> (t -> Maybe a) -> a infixl 5

Non-total application for partial transformers, such as normalTo.

Orphan instances

Bounded Double

Instance details

Methods

minBound :: Double

maxBound :: Double