Decorations

Classes

class Decor a where

The class of objects which could have decorations.

Minimal complete definition

Nothing

Methods

options :: a -> Options

Get the decoration data

setOptions :: Options -> a -> a

Set the decoration data

defaultOptions :: a -> Options

Default labeling settings

Types and records

data Options

Monoidal options list wrapper. Concatenates dually to a usual list.

data Option

Possible SVG options for a figure.

Constructors

Stroke String

Fill String

Thickness String

Dashing String

MultiStroke Int

LabelText String

LabelCorner (Int, Int)

LabelPosition CN

LabelOffset CN

LabelAngle Angular

SegmentMark Int

mkOptions :: [Option] -> Options

Puts a list of options to a wrapper.

getOptions :: Options -> [Option]

Extracts an option list from the Options wrapper.

newtype Decorated a

The transparent decoration wrapper for geometric objects. Inherits all properties of embedded object.

Constructors

Decorated (Options, a)

fromDecorated :: Decorated a -> a

The selector for the embedded object.

Decorators

newtype Decorator a

A wrapped decoration function with monoidal properties, corresponding to decoration options.

Constructors

Decorator (a -> Decorated a)

(#:) :: Decor a => a -> Decorator a -> Decorated a infixl 5

The infix operator for decorator application.

>>> aPoint # at (4, 5) #: label "A"
A:<Point (4.0, 5.0)>

>>> segment (4,5) (6,9) #: "s" <> dotted <> white
s:<Segment (4.0,5.0) (6.0,9.0)>

stroke :: Decor a => String -> Decorator a

The stroke color decorator.

white :: Decor a => Decorator a

The decorator for white lines.

fill :: Decor a => String -> Decorator a

The fill color decorator.

thickness :: Decor a => String -> Decorator a

The stroke-thickness decorator.

thin :: Decor a => Decorator a

The decorator for thin lines.

dashed :: Decor a => Decorator a

The decorator for dashed lines.

dotted :: Decor a => Decorator a

The decorator for dotted lines.

arcs :: Int -> Decorator Angle

The decorator for dotted lines.

label :: Decor a => String -> Decorator a

The decorator for labeling objects. Could be used as overloaded string.

>>> aPoint # at (4, 5) #: "A"
A:<Point (4, 5)>

loffs :: Decor a => CN -> Decorator a

The decorator for label offset.

lpos :: Decor a => CN -> Decorator a

The decorator for label position.

lparam :: (Curve a, Decor a) => Double -> Decorator a

The decorator for setting label on a curve at a given parameter value.