Figures

Constructors for geometric objects

Point constructors

origin :: Point

The origin point. Equivalent to aPoint.

aPoint :: Point

A point at the origin. Equivalent to origin.

aLabel :: Label

A label: the invisible point which coul be labeled using #: operator.

point :: XY -> Point

The constructor for a point with given coordinates.

point' :: Affine a => a -> Point

The generalized version of point.

pointOn :: Manifold a => a -> Double -> Point

The point on a given curve.

projectOn :: (Manifold c, Affine p) => p -> c -> Maybe Point

Returns a normal projection of the given point on the curve.

intersectionPoints :: (Curve a, Curve b, Intersections a b) => a -> b -> [Point]

Returns a list of intersection points as Point objects.

closestTo :: (Affine a, Affine b) => a -> [b] -> Maybe b

Returns a point from a list which is closest to a given one.

Line constructors

aLine :: Line

The template for a line.

aRay :: Ray

The template for a ray.

aSegment :: Segment

The template for a segment.

oX :: Line

The x-axis. Equivalent to aLine.

oY :: Line

The y-axis

line :: XY -> XY -> Line

line' :: (Affine a1, Affine a2) => a1 -> a2 -> Line

ray :: XY -> XY -> Ray

ray' :: (Affine a1, Affine a2) => a1 -> a2 -> Ray

segment :: XY -> XY -> Segment

segment' :: (Affine a1, Affine a2) => a1 -> a2 -> Segment

extendToLength :: Double -> Segment -> Segment

Returns a line, ray or a segment with given unit, in case of a segment -- with given length.

extendTo :: (Curve c, Intersections Ray c) => c -> Segment -> Maybe Segment

Returns a segment extended to a closest intersection point with a given curve.

normalSegment :: Curve c => c -> Double -> Maybe Segment

If possible, returns a line segment normal to the curve starting at a given parameter

heightTo :: (Affine p, Curve c, Intersections Ray c) => c -> p -> Maybe Segment

Returns a segment normal to a given curve starting at given point.

clipBy :: (Linear l, Intersections l c, Figure c, ClosedCurve c) => l -> c -> [Segment]

Returns a list of segments as a result of clipping the line by a closed curve.

Angle constructors

anAngle :: Direction -> Angle

The template for an angle with given value

angleBetween :: (Linear l1, Linear l2) => l1 -> l2 -> Angle

Returns the angle equal to the angle between thwo lines, located on the first one.

angleWithin :: (Affine a1, Affine a2, Affine a3) => a1 -> a2 -> a3 -> Angle

innerAngle :: Angle -> Angle

supplementary :: Angle -> Angle

vertical :: Angle -> Angle

reflex :: Angle -> Angle

bisectrisse :: Angular a => a -> Ray

Polygon constructors

aTriangle :: Triangle

The template for a triangle.

triangle2a :: Direction -> Direction -> Triangle

Returns a triangle with base 1 and two given angles.

aSquare :: Rectangle

The template for a square.

aRectangle :: Double -> Double -> Rectangle

The template for a Rectangle.

parametricPoly :: (Double -> XY) -> [Double] -> Polyline

Constructs a parametric graph as a Polyline.

polarPoly :: (Double -> Double) -> [Double] -> Polyline

Constructs a polar graph as a Polyline.

regularPoly :: Int -> Polygon

Constructs a regular polygon with given number of sides, enscribed in a unit circle.

Circle constructors

aCircle :: Circle

The template for a circle.

circle :: Double -> XY -> Circle

The constructor for a circle with given radius and coordinates.

circle' :: Affine a => Double -> a -> Circle

The generalized version of circle.

Misc

linearScale

Arguments

:: (Show a, Curve c)

=> (Double -> a)
labeling function

-> [Double]
range of the curve parameter

-> c
the curve

-> [Decorated Point]

Creates a scale as a list of labeled points on a given curve

modularScale :: (Trans c, Curve c) => Int -> c -> [Decorated Point]

Creates a circular integer scale, representing modular arithmetics.

Modificators

at :: Figure f => XY -> f -> f

Moves an object so that it's refPoint coinsides with a given one.

at' :: (Affine p, Figure f) => p -> f -> f

Generalized version of at transformer.

along :: (Figure f, Affine f) => Double -> f -> f

Rotates the figure which is Affine instance against it's refPoint so that it's refference angle (given by angle) councides with a given one.

along' :: (Figure f, Affine v, Affine f) => v -> f -> f

Generalized version of along transformer.

through :: Linear l => XY -> l -> l

Turns and extends the line so that it passes through a given point.

through' :: (Affine p, Linear l) => p -> l -> l

A generalized version of through.

translate :: Trans f => XY -> f -> f

Moves an object along given vector.

scaleAt :: Trans f => XY -> Double -> f -> f

Scales an object simmetrically (isotropically) against a given point.

scaleXAt :: Trans f => XY -> Double -> f -> f

Scales an object along x-axis against a given point.

scaleYAt :: Trans f => XY -> Double -> f -> f

Scales an object along y-axis against a given point.

on :: (Figure f, Affine f, Curve c) => c -> Double -> f -> f

Locates an affine object on a given curve at given parameter and aligns it along a tangent to a curve at this point.

normalTo :: (Curve c, Linear l) => c -> l -> Maybe l

If possible, turns the line so that it becomes normal to a given curve, pointing towards a curve.

flipAt :: Curve c => Double -> c -> c

Reflects the curve at a given parameter against the normal, if it exists, or does nothing otherwise.

vertexAngle :: PiecewiseLinear p => p -> Int -> Angle

height :: PiecewiseLinear p => p -> Int -> Segment

Orphan instances

WithOptions Segment

Instance details

Methods

options :: Segment -> Options

setOptions :: Options -> Segment -> Segment

defaultOptions :: Segment -> Options

WithOptions Ray

Instance details

Methods

options :: Ray -> Options

setOptions :: Options -> Ray -> Ray

defaultOptions :: Ray -> Options

WithOptions Line

Instance details

Methods

options :: Line -> Options

setOptions :: Options -> Line -> Line

defaultOptions :: Line -> Options

WithOptions Label

Instance details

Methods

options :: Label -> Options

setOptions :: Options -> Label -> Label

defaultOptions :: Label -> Options

WithOptions Point

Instance details

Methods

options :: Point -> Options

setOptions :: Options -> Point -> Point

defaultOptions :: Point -> Options

WithOptions Rectangle

Instance details

Methods

options :: Rectangle -> Options

setOptions :: Options -> Rectangle -> Rectangle

defaultOptions :: Rectangle -> Options

WithOptions Triangle

Instance details

Methods

options :: Triangle -> Options

setOptions :: Options -> Triangle -> Triangle

defaultOptions :: Triangle -> Options

WithOptions Polygon

Instance details

Methods

options :: Polygon -> Options

setOptions :: Options -> Polygon -> Polygon

defaultOptions :: Polygon -> Options

WithOptions Polyline

Instance details

Methods

options :: Polyline -> Options

setOptions :: Options -> Polyline -> Polyline

defaultOptions :: Polyline -> Options

WithOptions Circle

Instance details

Methods

options :: Circle -> Options

setOptions :: Options -> Circle -> Circle

defaultOptions :: Circle -> Options

WithOptions Angle

Instance details

Methods

options :: Angle -> Options

setOptions :: Options -> Angle -> Angle

defaultOptions :: Angle -> Options