Geometry

Reexports modules

module Base

module Point

module Line

module Circle

module Polygon

module Angle

module Decorations

Main interface

writeSVG :: (Figure a, SVGable a) => Int -> FilePath -> a -> IO ()

Creates SVG for a SVGable object and writes to a file with a given name.

showSVG :: (Figure a1, Integral a2, Show a2, SVGable a1) => a2 -> a1 -> Text

Creates a SVG contents for geometric objects.

(<+>) :: (Groupable a, Groupable b) => a -> b -> Group infixl 5

The appending operator for groupable objects.

group :: Groupable a => [a] -> Group

Returns a group of homogeneous list of objects.

Constructors for geometric objects

Point constructors

origin :: Point

The origin point

aPoint :: Point

The template for a point.

aLabel :: String -> Decorated Label

The template for a label.

point :: XY -> Point

The constructor for a point with given coordinates.

point' :: Affine a => a -> Point

The constructor for a point with given affine coordinates.

pointOn :: Curve a => a -> Double -> Point

The point on a given curve.

projectOn :: (Curve c, Affine p) => p -> c -> Point

Returns a point on the curve which is s normal projection of the given point on the curve.

Line constructors

aLine :: Line

The template for a line.

aRay :: Line

The template for a ray.

aSegment :: Line

The template for a segment.

oX :: Decorated Line

The x-axis

oY :: Decorated (Decorated Line)

The y-axis

line :: XY -> XY -> Line

line' :: (Affine a1, Affine a2) => a1 -> a2 -> Line

ray :: XY -> XY -> Line

ray' :: (Affine a1, Affine a2) => a1 -> a2 -> Line

segment :: XY -> XY -> Line

segment' :: (Affine a1, Affine a2) => a1 -> a2 -> Line

Angle constructors

anAngle :: Angular -> Angle

The template for an angle with given value

angleBetween :: (Linear l1, Linear l2) => l1 -> l2 -> Angle

Returns the angle equal to the angle between thwo lines, located on the first one.

adjacent :: Angle -> Angle

vertical :: Angle -> Angle

complementary :: Angle -> Angle

Polygon constructors

aTriangle :: Triangle

The template for a triangle.

triangle2a :: Angular -> Angular -> Polygon

Returns a triangle with base 1 and two given angles.

aSquare :: Polygon

The template for a square.

aRectangle :: Double -> Double -> Polygon

The template for a Rectangle.

parametricPoly :: (Double -> XY) -> [Double] -> Polygon

Constructs a parametric graph as a Polyline.

polarPoly :: (Double -> Double) -> [Double] -> Polygon

Constructs a polar graph as a Polyline.

regularPoly :: Int -> Polygon

Constructs a regular polygon with given number of sides, enscribed in a unit circle.

Circle constructors

aCircle :: Circle

The template for a circle.

circle :: Double -> XY -> Circle

circle' :: Affine a => Double -> a -> Circle

Modificators

at :: Figure a => XY -> a -> a

at' :: (Affine p, Figure a) => p -> a -> a

along :: (Figure a, Affine a) => Double -> a -> a

along' :: (Figure f, Affine v, Affine f) => v -> f -> f

through :: Linear l => XY -> l -> l

A coordinated version of through.

through' :: (Affine p, Linear l) => p -> l -> l

Turns and extends the line so that it passes through a given point.

on :: (Figure a, Affine a, Curve c) => c -> Double -> a -> a

Locates an affine object on a given curve agt given parameter and aligns it along a tangent to a curve.

normalTo :: (Curve c, Linear l) => c -> l -> l

Turns the line so that it becomes normal to a given curve, pointing towards a curve.