Line

Types

class (Figure l, Affine l, Curve l, Trans l) => IsLine l where

Class representing a linear object.

Minimal complete definition

refPoints

Methods

refPoints :: l -> (CN, CN)

asLine :: l -> Line

asRay :: l -> Ray

asSegment :: l -> Segment

newtype Line

The straight line, passing through two given points. The first point sets the Figure's refference point and a starting point of a line. The distance between refference points p1 and p2 sets the unit and internal scale, so that `p1 == l @<- 0` and `p2 == l @<- 1`.

Constructors

Line (CN, CN)

trivialLine :: Line

The trivial line with coinsiding refference points.

mkLine :: Affine a => (a, a) -> Line

The basic line constructor.

newtype Ray

Constructors

Ray (CN, CN)

mkRay :: Affine a => (a, a) -> Ray

The basic ray constructor.

newtype Segment

Constructors

Segment (CN, CN)

mkSegment :: Affine a => (a, a) -> Segment

The basic segment constructor.