Line

Types

class (Curve l, Affine l, Figure l) => Linear l where

Class representing a linear object: line, ray or segment.

Minimal complete definition

refPoints

Methods

refPoints :: l -> (CN, CN)

asLine :: l -> Line

asRay :: l -> Ray

asSegment :: l -> Segment

newtype Line

The straight line, passing through two given points. The first point sets the Figure's refference point and a starting point of a line. The distance between refference points p1 and p2 sets the unit and internal scale, so that `p1 == l @<- 0` and `p2 == l @<- 1`.

Constructors

Line (CN, CN)

trivialLine :: Line

The trivial line with coinsiding refference points.

mkLine :: (Affine p1, Affine p2) => (p1, p2) -> Line

The basic line constructor.

lineIntersection :: (Figure a1, Figure a2, Affine a1, Affine a2) => a1 -> a2 -> [Complex Double]

newtype Ray

The ray, passing through two given points. The first point sets the Figure's refference point and a starting point of a ray. The distance between refference points p1 and p2 sets the unit and internal scale, so that `p1 == l @<- 0` and `p2 == l @<- 1`.

Constructors

Ray (CN, CN)

mkRay :: (Affine p1, Affine p2) => (p1, p2) -> Ray

The basic ray constructor.

newtype Segment

The line segment, joining two given points. The first point sets the Figure's refference point and a starting point of a ray. The distance between refference points p1 and p2 sets the unit and internal scale, so that `p1 == l @<- 0` and `p2 == l @<- 1`.

Constructors

Segment (CN, CN)

mkSegment :: (Affine p1, Affine p2) => (p1, p2) -> Segment

The basic segment constructor.

end :: Segment -> CN

midPerpendicular :: Curve c => c -> Line