Line

Types

data Bounding

The type of line

Constructors

Unbound
a straight line

Semibound
a ray

Bound
a segment

data Line

The straight line of a given type, passing through two given points. The first point sets the Figures' refference point and a starting point of a line. The distance between refference points p1 and p2 sets the unit and internal scale, so that `p1 == l @<- 0` and `p2 == l @<- 1`.

Constructors

Line Bounding !(CN, CN)

class (Figure l, Affine l, Curve l, Trans l) => IsLine l where

Constrain for a linear object

Methods

bounding :: l -> Bounding

refPoints :: l -> (CN, CN)

Constructors

trivialLine :: Line

The trivial line with coinsiding refference points.

mkSegment :: (CN, CN) -> Line

The basic segment constructor

mkRay :: (CN, CN) -> Line

The basic ray constructor

mkLine :: (CN, CN) -> Line

The basic line constructor

Functions

extendAs :: Line -> Bounding -> Line

The extension of a line with the same refference points.

>>> mkSegment (0, 1) `extendAs` Semibound
<Ray (0.0,0.0), 0.0°>

>>> mkSegment (0, 1) `extendAs` Unbound
<Line (0.0,0.0), 0.0°>

>>> mkRay (0, 1) `extendAs` Unbound
<Line (0.0,0.0), 0.0°>

clipBy :: (IsLine l, Intersections l c, Curve c) => l -> c -> [Line]

Returns a list of segments as a result of clipping the line by a closed curve.